teste

+++ src/front_angular/src/app/core/chat/list-messages/list-messages.component.ts
@@ -23,22 +23,21 @@ export class ListMessagesComponent implements OnInit {
   id!: number | null;
   formMessage!: FormGroup;
   lengthMessages!: number;
+  
 
   constructor(
     private activatedRoute: ActivatedRoute,
-    
+
     private messagesService: MessagesService,
-    private channelPartageService: ChannelPartageService, 
-    private channelService: ChannelServiceComponent
-  ,
+    private channelPartageService: ChannelPartageService,
+    private channelService: ChannelServiceComponent,
     private fb: FormBuilder
   ) {}
 
   ngOnInit() {
-
-    //On récupère le channel 
+    //On récupère le channel
     this.channelPartageService.currentIdChannel.subscribe((id) => {
-      this.messagesChannel = []; //j'initialise les messages du channel vide 
+      this.messagesChannel = []; //j'initialise les messages du channel vide
 
       this.idChannel = id;
       console.log(this.idChannel);
@@ -55,20 +54,16 @@ export class ListMessagesComponent implements OnInit {
 
     console.log(this.channel);
 
-
-
     this.messagesService.getAllMessages().subscribe({
       next: (messages: Message[]) => {
-        messages.forEach(element => {
-          
+        messages.forEach((element) => {
           //je trie les éléments du channel
-          if(element.channel?.id == this.idChannel){
+          if (element.channel?.id == this.idChannel) {
             console.log(element);
-            
+
             //Je rajoute les éléments dans un nouveau tableau
             this.messagesChannel.push(element);
           }
-          
         });
         //console.log(messages[0])
         this.messagesList = messages;
@@ -96,13 +91,14 @@ export class ListMessagesComponent implements OnInit {
     }
   }
 
-  delete(id: number) {
-    this.messagesService.deleteMessage(id).subscribe((message) => {
-      console.log(message);
-      this.messagesService
-        .getAllMessages()
-        .subscribe((messages) => (this.messagesList = messages));
-    });
+  delete(id: number | undefined) {
+    if (id)
+      this.messagesService.deleteMessage(id).subscribe((message) => {
+        console.log(message);
+        this.messagesService
+          .getAllMessages()
+          .subscribe((messages) => (this.messagesList = messages));
+      });
   }
 
   save() {
